DNA
Prgoram written in Python to find and match people's DNA. It requires opening up two files, the database and the sequences. Then storing to memory using Python data structure, and compare the data base of name of people with a list of DNAs,  with the sequences itself, then decides whose DNA the sequences belong to.